safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/network/network-xhr-same-url-as-main-resource-expected.txt

CONSOLE MESSAGE: line 6: XHR loaded: 1
Tests that XHRs with the same url as a main resource have correct category.

Bug 68648
xhr
